+++ ScintSuite/SimulationCodes/FILDSIM/execution.py
@@ -317,7 +317,26 @@ def read_plate(filename):
 # -----------------------------------------------------------------------------
 # --- Energy definition FILDSIM
 # -----------------------------------------------------------------------------
-def get_energy(gyroradius, B: float, A: float = 2.01410178, Z: float = 1.0):
+def get_energy(gyroradius, B: float, A: float = 2.01410178, Z: float = 1.0,
+               relativistic: bool = False):
+    """
+    Calculate the energy given a gyroradius, FILDSIM criteria
+
+    Jose Rueda-Rueda: jruedaru@uci.edu
+    
+    :param  gyroradius: Larmor radius as taken from FILD strike map [in cm]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number (A=0 means electrons)
+    :param  Z: Ion charge [in e units] (if A=0, Z=1, electron charge)
+    :param  relativistic: Flag to use relativistic formula or not
+    """
+    if relativistic:
+        return get_energy_relativistic(gyroradius, B, A, Z)
+    else:
+        return get_energy_classic(gyroradius, B, A, Z)
+
+
+def get_energy_classic(gyroradius, B: float, A: float = 2.01410178, Z: float = 1.0):
     """
     Calculate the energy given a gyroradius, FILDSIM criteria
 
@@ -330,12 +349,56 @@ def get_energy(gyroradius, B: float, A: float = 2.01410178, Z: float = 1.0):
 
     :return E: the energy [in eV]
     """
+    # E[ev] = (1/2 * (r[cm]/100*Z*B[T])**2 / m) * ec
     m = ssp.amu2kg * A  # Mass of the ion
-    E = 0.5 * (gyroradius/100.0 * Z * B)**2 / m * ssp.ec
+    E = (0.5 * ((gyroradius/100.0) * Z * B)**2 / m) * ssp.ec
     return E
 
+def get_energy_relativistic(gyroradius, B: float, A: float = 2.01410178, 
+                            Z: float = 1.0):
+    """
+    Calculate the energy given a gyroradius, FILDSIM criteria
+    Relativistic correction, implemented to include runaway electrons
 
-def get_gyroradius(E, B: float, A: float = 2.01410178, Z: float = 1.0):
+    Alex Reyner: alereyvinn@alum.us.es
+
+    :param  gyroradius: Larmor radius as taken from FILD strike map [in cm]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number
+    :param  Z: Ion charge [in e units]
+
+    :return E: the energy [in eV]
+    """
+    if A == 0:
+        m = ssp.m_e_kg
+        Z = 1
+    else:
+        m = ssp.amu2kg * A
+
+    gamma = np.sqrt((gyroradius/100. * Z*ssp.ec *B /m)**2 /ssp.c**2 + 1) 
+    E = (gamma-1) *(m *ssp.c**2) / ssp.ec
+
+    return E
+
+def get_gyroradius(E:float, B: float, A: float = 2.01410178, Z: float = 1.0, 
+                   relativistic: bool = False):
+    """
+    Calculate the gyroradius given an energy, FILDSIM criteria
+
+    Jose Rueda-Rueda: jruedaru@uci.edu
+    
+    :param  energy: Energy [eV]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number (A=0 means electrons)
+    :param  Z: Ion charge [in e units] (if A=0, Z=1, electron charge)
+    :param  relativistic: Flag to use relativistic formula or not
+    """
+    if relativistic:
+        return get_gyroradius_relativistic(E, B, A, Z)
+    else:
+        return get_gyroradius_classic(E, B, A, Z)
+
+def get_gyroradius_classic(E, B: float, A: float = 2.01410178, Z: float = 1.0):
     """
     Calculate the gyroradius given an energy, FILDSIM criteria
 
@@ -348,6 +411,82 @@ def get_gyroradius(E, B: float, A: float = 2.01410178, Z: float = 1.0):
 
     :return r: Larmor radius as taken from FILD strike map [in cm]
     """
+    # r[cm] = sqrt(2*(E[eV]/ec)*m) / (Z*B[T]) * 100
     m = ssp.amu2kg * A  # Mass of the ion
-    r = 100. * np.sqrt(2.0 * E * m / ssp.ec) / Z / B
+    r = 100. * np.sqrt(2.0 * (E/ssp.ec) * m) / (Z*B)
     return r
+
+def get_gyroradius_relativistic(E, B: float, A: float = 2.01410178, 
+                                Z: float = 1.0):
+    """
+    Calculate the gyroradius given an energy, FILDSIM criteria
+    Relativistic correction, implemented to include runaway electrons
+
+    Alex Reyner: areyner@us.es
+
+    :param  energy: Energy [eV]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number. 0 means electrons
+    :param  Z: Ion charge [in e units]
+
+    :return r: Larmor radius as taken from FILD strike map [in cm]
+    """
+    if A == 0:
+        m = ssp.m_e_kg
+        Z = 1
+    else:
+        m = ssp.amu2kg * A
+
+    gamma = 1 + (E*ssp.ec)/(m*ssp.c**2)
+    beta = np.sqrt(1-1/gamma**2)
+    v = ssp.c * beta
+    r = 100. * (gamma*m*v) / (Z*ssp.ec) / B
+
+    return r
+
+def get_beta(E, B: float, A: float = 2.01410178, Z: float = 1.0):
+    """
+    Calculate the relativistic factor
+
+    Alex Reyner: areyner@us.es
+
+    :param  energy: Energy [eV]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number. 0 means electrons
+    :param  Z: Ion charge [in e units]
+
+    :return beta: v/c
+    """
+    if A == 0:
+        m = ssp.m_e_kg
+        Z = 1
+    else:
+        m = ssp.amu2kg * A
+
+    gamma = 1 + (E*ssp.ec)/(m*ssp.c**2)
+    beta = np.sqrt(1 - 1/gamma**2)
+
+    return beta
+
+def get_velocity(gyroradius, B: float = 1.9, A: float = 2.01410178, Z: float = 1.0):
+    """
+    Calculate the velocity given a gyroradius, FILDSIM criteria
+    Relativistic correction, implemented to include runaway electrons
+
+    Alex Reyner: alereyvinn@alum.us.es
+
+    :param  gyroradius: Larmor radius as taken from FILD strike map [in cm]
+    :param  energy: Energy [eV]
+    :param  B: Magnetic field, [in T]
+    :param  A: Ion mass number
+    :param  Z: Ion charge [in e units]
+
+    :return E: the energy [in eV]
+    """
+
+    E = get_energy(gyroradius,B,A,Z, True)
+    beta = get_beta(E,B,A,Z)
+    v = beta * ssp.c
+
+    return v
+