+++ ScintSuite/_Plotting/_settings.py
@@ -91,13 +91,17 @@ def plotSettings(plot_mode='software', usetex=False):
         'size': nml[mode]['axis_font_size']
     }
     mpl.rc('font', **opt)
-    mpl.rcParams['legend.fontsize'] = nml[mode]['legend_font_size']
-    mpl.rcParams['axes.titlesize'] = nml[mode]['title_font_size']
-    mpl.rcParams['axes.labelsize'] = nml[mode]['axis_font_size']
     try:
         mpl.rcParams['font.size'] = nml[mode]['inside_text_font_size']
     except KeyError:
-        mpl.rcParams['font.size'] = nml[mode]['axis_font_size'] - 2 # 2 points less than the axis label
+        mpl.rcParams['font.size'] = nml[mode]['axis_font_size']
+
+    mpl.rcParams['axes.titlesize'] = nml[mode]['title_font_size']
+    plt.rcParams['figure.titlesize'] = nml[mode]['title_font_size']
+    mpl.rcParams['axes.labelsize'] = nml[mode]['axis_font_size']
+    mpl.rcParams['xtick.labelsize'] = nml[mode]['tick_font_size']
+    mpl.rcParams['ytick.labelsize'] = nml[mode]['tick_font_size']
+    mpl.rcParams['legend.fontsize'] = nml[mode]['legend_font_size']
 
     mpl.rcParams['lines.linewidth'] = nml[mode]['line_width']
     mpl.rcParams['lines.markersize'] = nml[mode]['marker_size']
@@ -114,8 +118,8 @@ def plotSettings(plot_mode='software', usetex=False):
         mpl.rcParams['ytick.direction'] = nml[mode]['ytick_direction']
         mpl.rcParams['xtick.direction'] = nml[mode]['xtick_direction']
     except KeyError:
-        mpl.rcParams['ytick.direction'] = 'in'
-        mpl.rcParams['xtick.direction'] = 'in'
+        mpl.rcParams['ytick.direction'] = nml['default']['tick_direction']
+        mpl.rcParams['xtick.direction'] = nml['default']['tick_direction']
 
 
     # Print and return