+++ ScintSuite/version_suite.py
@@ -6,7 +6,7 @@ import numpy as np
 from ScintSuite._Machine import machine
 from ScintSuite._Paths import Path
 
-version = '1.4.4'
+version = '1.4.5'
 
 codename = 'Huki SX'